# small refator # handle delete simulation also on forntend site # if simulation is cancelled x appears
