+++ src/WrapperApp/components/Simulation/SimulationCard/SimulationCardHelpers.ts
@@ -1,0 +1,124 @@
+import { useSnackbar } from 'notistack';
+
+import { useDialog } from '../../../../services/DialogService';
+import { useLoader } from '../../../../services/LoaderService';
+import { FullSimulationData, useShSimulation } from '../../../../services/ShSimulatorService';
+import { titleToKebabCase } from '../../../../ThreeEditor/components/Dialog/CustomDialog';
+import { YaptideEditor } from '../../../../ThreeEditor/js/YaptideEditor';
+import { JobStatusData, SimulationInputFiles, StatusState } from '../../../../types/ResponseTypes';
+
+interface SimulationCardHelpersProps {
+	loadResults: ((jobId: string | null) => void) | undefined;
+	setDisableLoadJson: React.Dispatch<React.SetStateAction<boolean>>;
+	showInputFiles: ((inputFiles?: SimulationInputFiles | undefined) => void) | undefined;
+	simulationStatus: JobStatusData;
+	yaptideEditor: YaptideEditor | undefined;
+}
+
+const SimulationCardHelpers = ({
+	loadResults,
+	setDisableLoadJson,
+	showInputFiles,
+	simulationStatus,
+	yaptideEditor
+}: SimulationCardHelpersProps) => {
+	const { loadFromJson } = useLoader();
+	const { getJobLogs, getJobInputs, getFullSimulationData } = useShSimulation();
+	const { enqueueSnackbar } = useSnackbar();
+	const { open: openSaveFileDialog } = useDialog('saveFile');
+
+	const statusColor = (status?: StatusState) => {
+		switch (status) {
+			case StatusState.FAILED:
+				return 'error.main';
+			case StatusState.RUNNING:
+				return 'info.main';
+			case StatusState.COMPLETED:
+				return 'success.main';
+			case StatusState.LOCAL:
+				return 'warning.main';
+			case StatusState.CANCELED:
+				return 'common.black';
+			default:
+				return '';
+		}
+	};
+
+	const onClickLoadResults = () => {
+		loadResults?.call(null, simulationStatus.jobId);
+	};
+
+	const onClickGoToResults = () => {
+		loadResults?.call(null, null);
+	};
+
+	const onClickInputFiles = async () => {
+		const inputFiles = await getJobInputs(simulationStatus);
+
+		showInputFiles?.call(null, inputFiles?.input.inputFiles);
+	};
+
+	const onClickShowError = async (simulation: JobStatusData<StatusState.FAILED>) => {
+		const errorWindow = window.open();
+
+		const logfile = await getJobLogs(simulation);
+		const logfiles = Object.entries(logfile?.logfiles ?? {}).map(([key, value]) => {
+			return `<details><summary><h2 style="display: inline;">${key}</h2></summary><pre>${value}</pre></details>`;
+		});
+
+		if (!errorWindow) return console.error('Could not open new window');
+		errorWindow.document.open();
+		errorWindow.document.write(
+			`<html>
+			<head>
+				<title>Log Files</title>
+			</head>
+			<body>
+			<h1>Log Files</h1>
+			${logfiles}
+			</body>
+			</html>`
+		);
+		errorWindow.document.close();
+	};
+
+	const onClickSaveToFile = () => {
+		getFullSimulationData(simulationStatus)
+			.then((simulation: FullSimulationData | undefined) => {
+				if (yaptideEditor)
+					openSaveFileDialog({
+						name: `${titleToKebabCase(simulation?.title ?? 'simulation')}-result.json`,
+						results: simulation,
+						disableCheckbox: true,
+						yaptideEditor
+					});
+			})
+			.catch(() => {
+				enqueueSnackbar('Could not load simulation data', { variant: 'error' });
+			});
+	};
+
+	const onClickLoadToEditor = async (simulation: JobStatusData) => {
+		const inputJson = (await getJobInputs(simulation))?.input.inputJson;
+
+		if (!inputJson) {
+			setDisableLoadJson(true);
+
+			return enqueueSnackbar('Could not load json file', { variant: 'error' });
+		}
+
+		loadFromJson(inputJson);
+	};
+
+	return {
+		statusColor,
+		onClickLoadResults,
+		onClickGoToResults,
+		onClickInputFiles,
+		onClickShowError,
+		onClickSaveToFile,
+		onClickLoadToEditor
+	};
+};
+
+export default SimulationCardHelpers;
